requirements.txt
python-telegram-bot==21.5
anthropic==0.34.0
apscheduler==3.10.4
